requirements.txt
pandas
numpy
transformers
tensorflow-cpu
keras
scikit-learn
